# docs/openapi/project-service.yaml
openapi: 3.0.1
info:
  title: Project Service API
  version: 0.1.0
  description: Optional Project Management plugin for PlanCraftAI.
servers:
  - url: http://localhost:4005
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200': { description: OK }
  /v1/workspaces/{workspaceId}/plugins/project-management:
    get:
      summary: Get plugin settings
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Settings returned }
    put:
      summary: Update plugin settings
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectManagementEnabled: { type: boolean }
                sprintEnabled: { type: boolean }
                sprintLabel: { type: string }
      responses:
        '200': { description: Updated settings }
  /v1/projects:
    get:
      summary: List projects
      parameters:
        - in: query
          name: workspaceId
          required: true
          schema: { type: string }
        - in: query
          name: isActive
          schema: { type: boolean }
      responses:
        '200': { description: Projects listed }
    post:
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                workspaceId: { type: string }
                name: { type: string }
                description: { type: string }
      responses:
        '201': { description: Project created }
  /v1/projects/{projectId}:
    get:
      summary: Get project
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Project returned }
    patch:
      summary: Update project
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Project updated }
    delete:
      summary: Delete project
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Project removed }
  /v1/projects/{projectId}/statuses:
    get:
      summary: List statuses
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Status list }
    post:
      summary: Create status
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                order: { type: integer }
      responses:
        '201': { description: Status created }
  /v1/statuses/{statusId}:
    patch:
      summary: Update status
      parameters:
        - in: path
          name: statusId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Status updated }
    delete:
      summary: Delete status
      parameters:
        - in: path
          name: statusId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Status deleted }
  /v1/project-tasks:
    post:
      summary: Add task to project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectId: { type: string }
                taskId: { type: string }
                statusId: { type: string }
                sprintId: { type: string }
      responses:
        '201': { description: Mapping created }
  /v1/projects/{projectId}/tasks:
    get:
      summary: List tasks in project
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Project tasks returned }
  /v1/project-tasks/{projectId}/{taskId}:
    patch:
      summary: Update mapping
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
        - in: path
          name: taskId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Mapping updated }
    delete:
      summary: Remove mapping
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
        - in: path
          name: taskId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Mapping removed }
  /v1/projects/{projectId}/sprints:
    get:
      summary: List sprints
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Sprint list }
    post:
      summary: Create sprint
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                startDate: { type: string, format: date }
                endDate: { type: string, format: date }
      responses:
        '201': { description: Sprint created }
  /v1/sprints/{sprintId}:
    patch:
      summary: Update sprint
      parameters:
        - in: path
          name: sprintId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Sprint updated }
    delete:
      summary: Delete sprint
      parameters:
        - in: path
          name: sprintId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Sprint removed }
  /v1/projects/{projectId}/feedback:
    post:
      summary: Submit project feedback
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source: { type: string, enum: [client, link, email], default: client }
                text: { type: string }
                url: { type: string }
                author: { type: string }
              required: [text]
      responses:
        '201': { description: Feedback stored }
  /v1/projects/{projectId}/activity:
    get:
      summary: List recent project events (memory sink only)
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
        - in: query
          name: cursor
          description: Event id or timestamp cursor; returns events after this marker.
          schema: { type: string }
      responses:
        '200':
          description: Activity events returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer }
                        type: { type: string }
                        workspaceId: { type: string }
                        projectId: { type: string }
                        actor:
                          type: object
                          properties:
                            userId: { type: string }
                            displayName: { type: string }
                        entity:
                          type: object
                          properties:
                            type: { type: string }
                            id: { type: string }
                        timestamp: { type: string }
                        data: { type: object }
                  nextCursor: { type: string, nullable: true }
  /v1/projects/{id}/summary:
    get:
      summary: AI summary stub
      responses:
        '501': { description: Not implemented }
  /v1/sprints/{id}/health:
    get:
      summary: Sprint health stub
      responses:
        '501': { description: Not implemented }
  /v1/projects/{id}/stalled-tasks:
    get:
      summary: Stalled tasks stub
      responses:
        '501': { description: Not implemented }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas: {}
security:
  - bearerAuth: []